+++ spec/models/domain_spec.rb
@@ -94,6 +94,30 @@ describe Domain do
       @domain.registrant_update_confirmable?('123').should == false
     end
 
+    it 'should not find any domain pendings to clean' do
+      Domain.clean_expired_pendings.should == 0
+    end
+
+    it 'should not find any domains with wrong pendings' do
+      domain = Fabricate(:domain)
+      domain.registrant_verification_asked!('frame-str', '1')
+      domain.registrant_verification_asked_at = 30.days.ago
+      domain.save
+
+      Domain.clean_expired_pendings.should == 0
+    end
+
+    it 'should clean domain pendings' do
+      domain = Fabricate(:domain)
+      domain.registrant_verification_asked!('frame-str', '1')
+      domain.registrant_verification_asked_at = 30.days.ago
+      domain.pending_delete!
+
+      Domain.clean_expired_pendings.should == 1
+      domain.reload.pending_delete?.should == false
+      domain.pending_json.should == {}
+    end
+
     it 'should expire domains' do
       Domain.start_expire_period
       @domain.statuses.include?(DomainStatus::EXPIRED).should == false
@@ -180,6 +204,16 @@ describe Domain do
 
       @domain.statuses.count.should == 1
       @domain.force_delete_at.should be_nil
+    end
+
+    it 'should set expired status and update outzone_at and delete_at' do
+      domain = Fabricate(:domain)
+      domain.statuses.should == ['ok']
+      domain.set_expired
+      domain.changes.keys.should == ['statuses', 'outzone_at', 'delete_at']
+      domain.save
+
+      domain.statuses.should == ['expired']
     end
 
     it 'should know its create price' do