+++ app/api/repp/api.rb
@@ -5,10 +5,16 @@ module Repp
 
     http_basic do |username, password|
       @current_user ||= ApiUser.find_by(username: username, password: password)
+      if @current_user
+        true
+      else
+        error! I18n.t('api_user_not_found'), 401
+      end
     end
 
     before do
-      if request.ip != ENV['webclient_ip']
+      webclient_request = ENV['webclient_ips'].split(',').map(&:strip).include?(request.ip)
+      unless webclient_request
         error! I18n.t('ip_is_not_whitelisted'), 401 unless @current_user.registrar.api_ip_white?(request.ip)
       end
 
@@ -20,7 +26,7 @@ module Repp
       message = 'Certificate mismatch! Cert common name should be:'
       request_name = env['HTTP_SSL_CLIENT_S_DN_CN']
 
-      if request.ip == ENV['webclient_ip']
+      if webclient_request
         webclient_cert_name = ENV['webclient_cert_common_name'] || 'webclient'
         error! "Webclient #{message} #{webclient_cert_name}", 401 if webclient_cert_name != request_name
       else