+++ app/controllers/admin/bank_statements_controller.rb
@@ -16,7 +16,18 @@ class Admin::BankStatementsController < AdminController
   end
 
   def new
-    @bank_statement = BankStatement.new
+    @bank_statement = BankStatement.new(
+      bank_code: Setting.registry_bank_code,
+      iban: Setting.registry_iban
+    )
+    @invoice = Invoice.find_by(id: params[:invoice_id])
+    @bank_transaction = @bank_statement.bank_transactions.build(
+      description: @invoice.to_s,
+      sum: @invoice.sum,
+      reference_no: @invoice.reference_no,
+      paid_at: Time.zone.now.to_date,
+      currency: 'EUR'
+    ) if @invoice
   end
 
   def create
@@ -69,6 +80,8 @@ class Admin::BankStatementsController < AdminController
   end
 
   def bank_statement_params
-    params.require(:bank_statement).permit(:th6_file, :bank_code, :iban)
+    params.require(:bank_statement).permit(:th6_file, :bank_code, :iban, bank_transactions_attributes: [
+      :description, :sum, :currency, :reference_no, :paid_at
+    ])
   end
 end